requirements/requirements_prd.txt
-r requirements_base.txt

dj-database-url==0.2.2
dj-static==0.0.5
djangorestframework==2.3.12
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
